New Rouge Test Doc
The previous `asciidoctor-example.asciidoc` document now become just
an example document of how the syntax is rendered, all the tests and
elements details are now moved to a new `alan-syntax.asciidoc` doc.

The `sample.alan` contents are replaced with a sample adventure (no more
syntax tests), and it's used for the examples in the sample document.

Replace `asciidoctor-example.sh` with `build.sh`, a unified script that
converts all AsciiDoc sources to HTML and also compiles `sample.alan`.