Rouge Sass Themes
Begin setting up the Sass sources to generate custom CSS themes for
ALAN blocks, overriding the default theme. The CSS are still broken,
because they still use the old Highlight classes, and due to some
precedence problems which are preventing them from overriding the CSS
of the default Rouge fallback theme (created by Asciidoctor).

Tweak `custom-rouge-adapter.rb` to add some extra attributes to the
HTML code for highlighted blocks, in order to be able to target themes
on a per-language basis.