Update Rouge ALAN Sample
Add block comment at the beginning of the source, to test it with the
lexer and themes.

NOTE: Due to a bug in ALAN 3.0beta8 build 2207, I had to add an empty
line at the beginning of the source, and use `LF` EOL instead of Win's
native `CRLF`, otherwise it wouldn't compile (see alan-if/alan#30).