Rouge Sass: Integrate Upstream Updates
Tweak the Sass sources for Rouge in order to use the new var names of
the default ALAN theme, so that it will be easier to integrate Rouge
specific Sass sources in ALAN Docs.

Tweak the ALAN Rouge Sass template to mirror its Highlight counterpart:
same colours and styles of the default theme are assigned to tokens,
even if this means hiding the extra syntax elements of Rouge. But right
now we need to add Rouge support in ALAN Docs, so we don't have time to
play around with new palette colours.